print: initial attempt to add a print command to the program code
TEST: verify with the input stream set to PRINT, that upon calling the
command parser, that the program word is added to the code, and that the
code added is the code for PRINT.

Quite a bit of plumbing that will be needed was implemented.  This was a
little beyond "just making the test pass" but there was no reason on
adding code that will be replaced later.  This included:

A new command code class derived from code, which takes a keyword in the
constructor.  This keyword is added to a static map of the keyword to a
pointer to a command code instance.  There is also a static find code
function for looking up a keyword (of a command) in this map and
returning a pointer to the command code instance.  This class also has a
pure virtual function for parsing the command.

Functionality was added to the command parser parse function including
a check if there is nothing in the input stream (blank line), needed to
keep the first test passing; getting the code for the "PRINT" keyword
(a hack, should be read from the input stream, but makes the test pass);
and calling the parse from for the command.

A new print source file was added that contains a print code class
derived from the command code class, which passes the "PRINT" keyword
to the command code constructor; a print code instance; and the parse
function, which puts the print code into the program code.

There is some missing code, like checks for failures that were not
needed for the tests implemented so far.

OTHER: The names of the constant codes did not follow the snake casing
convention and were therefore renamed to conform.