command: parse another (END) command
TEST: verify that a different command is properly parsed and encoded
into the program code.

The previous change for parsing a simple PRINT command hard-coded the
PRINT keyword, so obviously trying to parse a different command was
going to fail.

To make this test pass, a new source file for the END command was added,
which is essentially a clone of the PRINT source file.  Both of these
commands' parse functions just put their code into the program line.

The command parser parse function was modified to call a new parse
keyword function that scans for alpha characters adding them to a string
to be returned when the first non-alpha character is found.  This
keyword string is then passed to the command code find function.

OTHER: renamed the command code find code function to just find (the
extra 'code' was redundant).