command: check for an error if no alphabetic keyword is present
TEST: verify an error is throw when a keyword not starting with a letter
is first in the input stream.

A check for an empty keyword was added after calling the parse keyword
function.  The error message won't be checked at this time since the
message added will need to be changed later (since eventually an
assignment may be present without the LET keyword).