print: parse a statement with an integer constant
TEST: verify that the single expression PRINT statement is parsed
correctly.

The first change was to skip possible white space after the PRINT
keyword, which was added for the parse keyword function before
returning.

The print code parse function was modified to first check if at the end
of the line (for a blank PRINT statement).  Otherwise a expression
parser instance is created and called (which adds the constant integer
to the code with its operand), then the new print integer code is added
to the code line, followed by the PRINT command code.  Just the PRINT
command code is added if the input stream was at the end of the
statement.

OTHER: rearranges constant number parser implementation functions to
the proper order (process input should follow that parse function that
calls it).