print: parse a statement with a double constant
TSET: verify that the single expression PRINT statement with a double
constant is parsed correctly.

A new print double code was added.  The return data type of the
expression parser is now captured and checked, it double a print double
code is added to the program line, otherwise an print integer code is
added (these are the only data types returned from the expression
parser at this time).