const-num: combined the double and integer dictionaries
There needs to be just one dictionary for the double and integer
constants because at run time either type of constant could be used as
the other type (except for double constants that out of the range of an
integer).

Also merged the test require operand macros into a single macro with an
additional argument for the data type.