command: allow case insensitive command keywords
TEST: check that the case of command keyword are compiled correctly.

This test initially failed because only the upper case form of the
commands were found in the command code keyword map.  A new case
insensitive string class was added (see below) and the unordered map was
changed to a regular map using this case insensitive string as the key
type (an unordered map would have required a hash function, and nothing
significant would have been gained using one of these as the map size
will not be that large).

The command code find function argument was changed to a case
insensitive string, and the compiler get keyword was changed to return
a case insensitive string.

The case insensitive string class is based on standard strings with a
case insensitive character traits derived from the standard character
traits overloading functions required to make it case insensitive.
Overloads of the relation operators allowing case insensitive strings to
be compared to standard strings (either left hand size of right hand
side of the operators).  Overloads of the stream inserter and extactor
functions were also added.  A complete set of unit tests were added for
this new class and all of these function overloads.