const-num: implemented constant double recreate function
TEST: verify that the constant number for a double constant instruction
added to program code is created correctly.

To get this test to compile and run, quite a bit of infrastructure
needed to be implemented or modified, including:

1) A add code line function to the program unit class to add a line of
code (generated by the compiler) to the end of the program code.  A code
member was added to the program unit class.

2) A recreate virtual function to the code class.  For now, a blank
dummy recreate function was added to the code base class so that all the
recreate functions didn't need to be added right away (note that this
function does not get covered by the test, and will be removed as soon
as all the code recreate functions are implemented.

3) A constant number code class that contains the recreate function
implementation for constant number codes.  Moved the constant double and
integer code instances from the constant number compiler source file to
a new constant number code source file (and changed their types to the
new class).

4) A program reader class for reading the program code, which is
basically a smart iterator.  The only function implemented is for
getting an instruction code pointer and an operand.  The create program
reader function was added to the program unit class.

5) A recreator class, which contains a reference to the program unit
(with the code to recreate), a program reader instance for reading the
program code, and a stack for holding the recreated string parts of the
statement (a private stack item structure).  Member functions include
recreate one code, get constant number operand, push a string to the
stack, and a top to get the string on top of the stack.

6) The static count member of the code class was replaced with a vector
for holding the pointers to the codes.  This is needed so that the code
instance can be obtained for a code value contained in the program.  A
new static private add code function was added to add the pointer of
each code to this vector, where the size of the vector before the add is
the value that the code will be set to.  A new static get code function
was added so that a code instance pointer can be obtained for a code
value.

7) The program word instruction code was changed to return a pointer to
a code so that it can be used to access the virtual functions of the
code.  A begin function was added to the program code class for getting
an iterator to the beginning of the code vector.

OTHER: added some missing spaces in constant number unit tests source
file, made the get value function in the code class constant.