command: recreate a blank print statement
TEST: verify that a blank PRINT statement (which consists of just a
single PRINT code) is recreated correctly.

While the test could have been made to pass simply by have the print
code recreate function push a "PRINT" to the recreator stack, the
complete code was implemented, where the new print recreate function
gets the keyword for the command to push.

The new get keyword function was added to the intermediate command code
base class.  This required a new static map in the command code class
for taking a command code value and returning the command name string.
The current command names map was renamed to command codes since it is a
map from a keyword and returning a pointer to command code instance.
The new map (maps cannot be searched for the key value) maps from a code
value (simpler then a pointer) to the command keyword (a simple c-string
pointer).  The command code constructor was modified to also add to this
new map.

OTHER: corrected the class name on external statement for the constant
double code for another command unit test (somehow that compiler was
accepting the wrong type), changed the tag on the first command unit
test case, and renamed the code line variable in the constant number
recreate unit test.