command: recreate an end command
TEST: verify that an END statement (which consists of just an END code)
is recreated correctly.

This test was made to pass simply be implementing the end code class
recreate function that pushes the end keyword to the recreator stack.

This is identical to the print code recreate function, so the function
could have been pushed up to the command code intermediate base class,
however, the print recreate function will need to do more work.  If it
turns out that more commands just output a keyword, then at that time
this functionality will be pushed to the command code class.