command: recreate a print with constant statement
TEST: verify that a PRINT with a constant is recreated correctly.

Access to the recreator was moved into the program unit class with a new
recreate line function that returns the recreated line requested (by a
line index).

The program unit was given a private line info structure for containing
the offset (into the code vector) and size of each line appended to the
code vector.  The append code line function (renamed from add) now adds
the line info for the code line being added to a new line info vector
member before adding the code line to the code vector.

The program reader class was given an end iterator, which is initialized
in the constructor using the new offset and size arguments.  A has more
code access function returns if there is more code to read (if the
iterator is not equal to the end iterator.  The program unit create
program reader function was modified to obtain the info for the line
index requested to create the program reader.

The recreator was given several new functions including a function
operator function to recreate an entire line, a pop function to pop the
top string on the stack, a top add space function to add a space
character to the top string, a top add function to add a string to the
top string, and an empty function to return if the stack is empty.  The
program reader instance is now passed into the constructor instead of
being created in it.

The recreator function operator function loops while the program reader
has more code calling the recreate one code function (which is now
private), and returning the top string when done.  The new program unit
recreate line function creates the recreator instance and calls this
function.

All unit tests creating a recreator instance was modified to use the new
program recreate line function.