recreate: refactored code recreate function mechanism
A bad pattern was starting develop where the number of code classes was
starting to explode (as they did on the IBCP).  In addition, with the
code virtual recreate function, a code instance is no longer a simple
single 16-bit integer (which is desirable because it can be passed
around by value, but with virtual functions, it will also have a vtbl
that will be part of the instance).

As an alternative, the recreate functions were made separate functions
not associated with a class.  They wouldn't be sharing any variables
with the other functions (compile, execute) except the code value.
These recreate functions are registered when the code instance is
created - the code base class now contains a vector of recreate
functions accessed by code value (an index).

The recreate function is passed through to the code class constructor
and is added to this vector.  A recreate function was added to the code
class for calling the recreate function of the code, which is not
virtual function.  The standard function template class is used for
this, so function operator classes (or lambda expressions) could be used
instead of a function.

The constant number code class was eliminated and its header file
removed.  Since the recreate functions are no longer members of the code
classes, the code instance needs to be referenced to access the command
keyword functions.  A blank print item recreate function also needed to
be added.

OTHER: removed the explicit code with value constructor declaration as
the function was previously removed.