compile: refactored command code compile function mechanism
Repeating the same changes as for the recreate functions, the command
code virtual compile function was eliminated.  However, instead of using
a vector to store the compile functions, a map was used instead since
not all codes are command codes.

For self documentation purposes (mainly for the command code maps), a
type alias for uint16_t was added to the code class.  This better
explains (with code) what is being mapped.

Accessing the command code maps was changed from using the find
function, which was immediately used to get the second value of the
iterator (the value associated with the key) to using the bracket
operator.  There was a problem the command code keyword look up in that
it didn't check if the keyword existed before getting the code pointer
(getting the second value of the end iterator would cause a segfault).
With the bracket operator, it will at least added a new entry with a
null pointer and return that (not ideal, but good enough for now).

In the command unit tests, the find code look ups for the print and end
codes was replaced with simple external declarations (the find was used
to the print and end classes didn't need to be exposed to the world, but
those classes have now been eliminated.

OTHER: changed the code instances to using the C++11 initializer syntax.