project: modified to build a single shared library for unit tests
The way the cmake build file was configured, it was building each source
file for each of the four unit tests.  This obviously was not the right
way, so a shared library was created with all of the core files, and
this library is linked into each of the unit test executables.