compiler: compile a multiple line code into the program
TEST: given a string with multiple lines, verify that each line is
compiled into the program - verification is accomplished by requesting
the program recreate each source line and checking for the expected
statements.

A new compile source function was added to program unit class, which
takes an input stream.  Each line of the input stream is read, and the
line is compiled and appended to the program.

The resulting code in this function is a bit tedious and verbose - the
same pattern exists through the unit tests, so a bit a refactoring is
needed.

OTHER: rearranges the program unit class member functions a bit; made
use of the bracket operator in the command code constructor when adding
to the maps (this is a bit cleaner looking then using the emplace
function).