compiler: changed input from input stream to string (refactor)
This is the first refactoring to clean up calling the compiler.

All users that created a compiler instance passed in a reference to a
temporary input string stream (created from a string), the input string
stream was made a member of the compiler class (replacing the input
stream reference.  The constructor argument was changed from an input
stream reference to a string reference.

Since the stream is now a member of the compiler class, the constant
modifier needed to be removed from the various member functions that
modify the input stream.

This change made another problem evident in the expression compiler
implementation class, which did not have the compiler member as a
reference.

The tests that checked that an error was thrown and then checked the
error message and column were put into sections so that setup of the
tests did not need to be repeated.  There were two tests that modified
the input stream before calling a compile routine so that the error
column could be checked when the number is not at the beginning of the
stream.  This could no longer be done so these tests were replaced with
a new command unit test with the error in a print statement.