compiler: put code line instance into compiler class
This is the second refactoring to clean up calling the compiler.

All users of the compiler class passed in a reference to a code line,
which always started empty.  So the code line instance was put into the
compiler class and a function was added to get the code line from the
compiler (this function moves the code line output of the compiler
instance).  The constant modifier needed to be removed from the add
instruction functions since they are no modifying a member variable.