compiler: modified command compiler to create a compiler instance
This is the final refactoring to clean up calling the compiler.

A compiler instance was created just before creating and calling the
command compiler function operator function instance, so it was modified
to create and hold the compile instance itself.  The constructor
arguments were changed from a compiler reference to a source line
string reference and program unit reference so that it (the underlying
implementation class) can create the compiler instance.

This allowed the call to be reduced to a single line (by creating the
instance and calling the function operator function).  In total, this
one line replaces the original five statements lines (create input
string stream instance, create a program code line instance, create a
compiler instance, create a command compiler instance, and call the
command compiler instance function operator function.  And the net
change of lines in the other classes was zero to accomplish this.