recreator: cleaned up the code a bit (refactor)
Moved the functionality from the print recreate function that prepended
the command keyword to the beginning of the top string on the stack
(containing a fair bit of knowledge of the recreator class internals)
to a new recreator prepend keyword function (and the number of
statements was reduced from five to four be making use of the swap
function to exchange the keyword and the top string).

Renamed a couple of the other recreator functions and added a push
keyword function to eliminate calling the get keyword function from the
callers (this function will be used for many of the other commands, if
not all of them).