const-num: implement basis for executing number constants
TEST: compile an integer constant, create an executer instance, execute
one code (the constant), and check that the top of the execution stack
has the number.

The executor is going to need a bunch of infrastructure.  Instead of
trying to build all that at once, the TDD approach will by used here to
slowly build this infrastructure where just enough code will be
implemented (though possibly invalid) to make each test pass.

A new empty executor class (source and header file) was added.  An
executer class could now be instanced.  An empty execute one code
function was added.  This function could now be called.

Jumping ahead a tiny bit, the execution stack will be a union of the
various value types that could be on the stack (integer, double, string,
variable references, etc. - the execute functions will know exactly what
types will be on the stack).  A stack item union was added to the
executer class with a single integer variable.

Finally a top function was added to the executer class that returns a
stack item (this will probably need to return a reference, or there may
be separate top functions for each type).  To make the test pass, a
local stack item with its integer value hard-coded to the expected value
of the test.