executer: modified program unit class to create an executor instance
The executer is going to need access to the program unit for the code
vector and information from the dictionaries (constant values, etc.).
For speed, the executor will use plain arrays for this data (a standard
vector is accessed as a plain vector).  It would be inefficient to
access this info through member functions of the program unit class
plus this would exposes the program unit internals unnecessarily.
Better would be to give this info to the executer class, and since the
program unit has this info, a create executor function was added to the
program unit class.  The unit test was modified to call this function.

OTHER: the create program reader function was made private to the
program unit class since it was the only caller.