program: introduced new word type
Previously the program word type was defined in the Code class as just
the type.  Since this type applies to all program words, the type type
was removed from the Code class and a new word type type was added in
its own header file (needed by the separate code and program word header
files).  All of the uses of the unsigned 16-bit integer type was
replaced with this new word type.

Doing this allows other program word types to be tried.  It may turn out
that using a 32-bit unsigned integer will produce faster run-times than
using a 16-bit as it saves from having to convert to a longer word in
the processor.  For now a 16-bit word size will be used as it produces a
more compact program.