executer: implemented full support for constant number execution
TEST: verify that the execution two different integer constants work
correctly.

This test initially failed because the execution code was simply
returning the value of the first test (to just make the test pass).
To make this test required the constant number dictionary to also hold
the numeric (integer) values (with the string values).  Since not all
dictionaries need this, a new constant number dictionary class was added
derived from the dictionary class containing a vector of integer values.

The add function for the constant number dictionary class now handles
determining the constant number code needed (has a floating point flag
argument in addition to the number string).  The number string is
converted to an integer (converting doubles were not needed to make the
test pass) and added to the integer values vector.  The base dictionary
add function is called to generate the operand index.  This function
returns a new constant number info structure containing the code value,
operand index and data type.

The compiler add constant number instruction now calls the constant
number dictionary add function and its code argument was replaced with
the floating point flag.  The constant number compile function was
modified to just have call to this function and passes to it its
floating point flag.  Unit test callers of this compiler functions were
updated to pass a true or false value depending on which code it
desired to be added to the code.

In order for the executer top function to return the correct integer
value, a pointer to the beginning of the program code and a pointer to
the constant number dictionary integer values vector data are passed to
the executer class instance through its constructor and put into new
members for these, adding the access functions necessary to get these
pointers.

The top function was modified to get the operand from the second word in
the code, the operand index, which is used to get the value out of the
constant integer values.  (This is not the correct implementation, but
it was the minimum to make the test pass.)