const-num: added program counter to executer class
TEST: verify that two integer codes can be executed consecutively.

Previously the top function simply grabbed the operand from the second
code word, so this test showed this is no longer sufficient.  A program
counter was needed to step through the program code.  To make this test
pass, the execute one code increments the program counter (over the
code), and the top function gets the operand from the program counter
and then increments it.