executer: added evaluation stack to executer
TEST: verify after execute two consecutive integer codes that the value
on top of the stack is the second integer, and after popping this value
off of the stack, the first integer is on top of the stack.

To make this test pass, a stack was added to the executer class.  In
order to be able push a stack item to the stack by value, the stack
item union was changed to a structure with the union so that
constructors can be added for values.

An integer value constructor was added.  The top function was changed to
return a reference to the top item on the stack (and moved to the header
as an in line function).  A pop function was added as an in line
function.  These functions are in line to allow the faster possible
execution of the program.

The execute one code function was changed to also push the integer
operand to the stack.  This is still not the correct implementation
(assumes only integer constant codes), but for now it makes the current
test pass.