executer: added support for double constants
TEST: verify that a single double constant code can be executed and that
its value is on top of the stack.

To make this test compile, a double value needed to be added to the
stack item along with its constructor.  To make this test pass, the
execute one code was modified to look at the code word of the
instruction (previously an integer constant was assumed).  If the code
is the same as the constant double code, then the expected value is
pushed to the stack, and though this is obviously not correct, it is
sufficient to make this test pass.