executer: allow different double constants
TEST: verify that the execution of two different double constants work.

To make the previous test pass, the expected value was hard-coded into
the execute one code function.  The goal was to add double value support
to executer stack.  To get other double constants to work, double value
support was added to the constant number dictionary.

Conversion of double number strings to double values was added to the
constant number dictionary add function along with a vector of double
values member.  Both a double and integer value needs to be to the
values vector since a constant can be used as either.

The executer class gained a pointer to the constant double values (also
added to its constructor).  The execute one code function was rearranged
to get the operand after getting the code and modified to push a
constant double value to the stack.