executer: implement execute functions (refactor)
The execute one code function was growing and would continue to grow as
more codes were implemented.  This implementation was temporary to get
the current tests to pass, but it was known that all codes would have
their own functions for execution.

An execute function pointer alias (taking a reference to the executer
instance and returning nothing) was added and a static vector of execute
function pointers was added to the code class.  The code constructor
with newly added execute function pointer now adds the execute function
pointer for the code to this vector.  An access function to this
vector's array was added, which is used to pass this array to the
executer during its construction (to a new member).

A simple function pointer was used because it turns out that the
standard function template is rather heavy and is four times the size
of a function pointer.  A function pointer is sufficient, so for
efficiency during execution, this is all that is necessary.

For the unimplemented execution functions (print items, print and end),
the execute function pointer is initialized to a null pointer.

The execute one code function was modified to call the execute function
by pointer by indexing this array by the code value and passing a
reference (to the executer itself) to the function.

Execute functions were implemented for the constant double and integer
codes, which get their operand (a new get operand access function), and
push their value (new push constant double and integer access functions)
to the stack.  The access function were made in line functions for
execution efficiency.