const-num: reuse dictionary entry for same constant string
TEST: verify that when the same constant string is used multiple times
in the code that the dictionary entry is reused.

Previously if the same numeric constant string appeared in the code, the
map entry would be reused, but the new entry was added to the value
vector in the constant number dictionary.  The result being that the
wrong value of the constant would be pushed to the stack.

To make this test pass, the constant number dictionary add function must
detect when a new entry was added to the base dictionary (return a new
index operand) or existed (return the previous index operand).

The base dictionary add function was modified to return a new structure
containing the index operand and a flag if the entry already exists.
Privately, the add function was refactored into two sub-functions to
better clarify what it is doing.  A private key map entry structure was
adding containing the key map iterator and if the key exists flag.