const-num: check for integer values to large for an integer
TEST: verify that a large integer number string (with no decimal point
or exponent) is compiled as an double constant.

To make this test pass, a try/catch block was added around the standard
string to integer function call, which throws an out-of-range exception
if the value cannot be represented as an integer.  If this exceptions is
thrown, the floating point flag is set and the number string is
converted to a double and a double constant is created.

There are two resulting issues.  An integer value is still created
(which the compile lets slide because doubles implicitly convert to
integers regardless of whether the double can be represented as a
integer).  The size of the constant number dictionary add functions is
getting rather unwieldy long and needs to be refactored.

Both of these issues are will be handled later as they are not required
to make the test pass, and there is more that needs to be added.