const-num: refactored dictionary add function
The constant number dictionary add function was getting large and
complicated.  With the unit tests in place, this function was refactored
into a sub-function and a private constant number converted class.  A
to do command was added that a double value may not be able to be
converted into an integer.  The constant number info structure was
renamed to the constant number code info structure to better clarify
its contents.