const-num: check for double values to large for a double
TEST: verify that a large double number string causes a compile error to
be thrown.  Also check the error message and column.

A try/catch block was added to the constant number compiler
implementation compile function when adding a constant number
instruction for a standard out of range exception.  A compile error is
thrown.

The column of the error is also verified, and to make this test pass, a
column of zero is passed to the compile error constructor.  The current
column at this point is set to the last character of the constant and
therefore contains the wrong column.