const-num: throw correct column for out of range error
TEST: verify that when the too large double constant is not at the
beginning of the input stream that the column of the error is correct.

To make the last test pass, the error column was simply set to zero.
The problem with the current column was that it was set to the last
character of the too large constant.  The error needs the first column
of the number.

To make this test pass, a first column member was added to the constant
number compiler implementation class, which is initialized to the
current column of the compiler before number parsing is started.  This
first column is then passed to the compile error constructor when
throwing an error for an out of range number constant.

This unit test was added to the command unit tests, where an actual
PRINT statement with too large constant was the test input.  This puts
the starting column of the constant away from the beginning of the input
stream so that error column can be tested.

OTHER: corrected a previously undetected mistake with the unsigned type
used for columns.  The actual type from the stream tell get position is
a stream position type (probably a signed long type).  This function
could return a -1 for an error.  The character type was used for the
compile get column function, which is wrong (even though this is a
signed integer type, is probably not large enough).  All column types
were changed to a plain integer (there should be no columns larger
then will fit in an integer).