command: throw correct error column with invalid command
TEST: verify the correct column reported with the error when an invalid
command is found.

The column reported was not taking into account the actual column of the
bad command and was assuming that it was at the beginning of the input
stream.  This was corrected by storing the column after skipping white
space in the compiler get keyword function into a new keyword column
member.  An access function was added for this value.

The command compiler implementation compile function was modified to
get this keyword column when throwing the error (previously was just
return a zero).

The unit test was extended to check both the error message and the error
column.