compiler: added length to compile error
TEST: verify correct length is throw with compile error.  This test was
added to all location where column was also being checked.  In most
cases, the expected length is 1, however, for number out of range
errors, the length should be the length of the number string.

To make this test pass, a length member was added to the compile error
structure, and a length argument was added to the constructor and was
given a default value of 1 (most constant number parsing errors are of
length one).  The throw of the out of range error was modified to also
pass the length of the number string.