executer: run an entire program
TEST: verify that an entire program can be run.

The program test in the command unit tests was extended to execute the
program after the recreate program test using the same program.

A run function was added to the program unit, which creates an executer,
calls the executer run function within a try block.  This is to catch
run time errors (of which none exists yet), but also to catch the end of
program exception (which is a normal end of program, where an exception
is used so no check needs to be performed during execution).

The executer run function first resets the program counter (by calling
a new reset function also called from the constructor), and then enters
an infinite loop calling the execute one code function repeatedly.  (As
mentioned until an exception is thrown for the end of the program or a
run time error).