executer: make sure program doesn't run past the end of the program
TEST: verify that a program without an END command terminates properly.

An END is not required in a BASIC program.  To make sure the executer
doesn't run past the end of the code, the program unit run function was
modified to add a temporary end code to the end of the code before
running the program, which is then removed afterward.

This was accomplished with a local program end guard class.  Its
constructor, which takes a reference to the code, adds an end code to
the end of the code vector.  Its destructor then removes this end code
(using a new pop back function added to the program code class).

Using this class insures that the end code is removed regardless how the
program terminates (normally or a run time error), without an complex
code in the run function.