program: recreate an entire program
TEST: verify that an entire program is recreated (and also test that the
temporary end code is not output as an END at the end of the program).

This unit was added to the last one (which was rearranged a bit), with
the recreate test after the execute test (after the automatic end code
is added and removed from the program code).

A recreate function was added to the program unit class, which just
repeated calls the recreate line function for each line in the program
(the size of the line info vector).