compiler: detect errors when compiling source code
TEST: verify that if there are errors in the source code, that a failure
status is returned.

The program compile source function was modified to return a success
status.  The previous unit test was modified to verify that a success
status is returned.  The new unit test attempts to compile source code
containing errors to verify a failure status is returned.

Inside this function, the compile line call was put into a try/catch
block.  If a compile error is thrown, the error is counted.  Upon
return, a success returned only if the error count is zero.