compiler: report errors when compiling source code
TEST: verify that error messages are reported on lines with errors.

This test is in addition to the last test.  An error message format
similar to GCC was selected.

A reference to an output stream was added to the compile source function
of the program unit class so that errors can be reported.  When a
compile error is caught, the new handle error function is called, which
appends an empty code line (needed to keep the line info vector
correct), the number of the line is obtained, and the new output error
function is called.

All callers of the compile source function was updated for the new
output stream reference argument (in all cases, the output string stream
is unused).